base fixes enabled/required by C99 (1 / Type handling cleanup)
(all changes that don't have anything to do with [u]int*_t,
Num-introduction, float-unboxing, autoconf, or the new math builtins --wrog)

was:
- use C99-like macros for printf
- use [u]int*_t types (autoconf can verify they exist)
- abstract Num and Objid out
- default to 64 bit integers
- un-pointerify floating-point numbers (makes no sense on modern hardware)

+++ utils.c
@@ -140,7 +140,7 @@ complex_free_var(Var v)
 {
     int i;
 
-    switch ((int) v.type) {
+    switch (v.type) {
     case TYPE_STR:
 	if (v.v.str)
 	    free_str(v.v.str);
@@ -164,7 +164,7 @@ complex_free_var(Var v)
 Var
 complex_var_ref(Var v)
 {
-    switch ((int) v.type) {
+    switch (v.type) {
     case TYPE_STR:
 	addref(v.v.str);
 	break;
@@ -184,7 +184,7 @@ complex_var_dup(Var v)
     int i;
     Var newlist;
 
-    switch ((int) v.type) {
+    switch (v.type) {
     case TYPE_STR:
 	v.v.str = str_dup(v.v.str);
 	break;
@@ -208,7 +208,7 @@ complex_var_dup(Var v)
 int
 var_refcount(Var v)
 {
-    switch ((int) v.type) {
+    switch (v.type) {
     case TYPE_STR:
 	return refcount(v.v.str);
 	break;
